requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
bcrypt==4.1.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.114.2
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.1.0
h11==0.14.0
idna==3.10
makefun==1.15.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pwdlib==0.2.0
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
